// nft-drop-starter-project-final-master/app-backup/src/CandyMachine/interfaces.ts
import * as anchor from "@project-serum/anchor";

export interface ICandyMachineState {
  //   authority: anchor.web3.PublicKey;
  itemsAvailable: number;
  itemsRedeemed: number;
  itemsRemaining: number;
  treasury: anchor.web3.PublicKey;
  tokenMint: null | anchor.web3.PublicKey;
  isSoldOut: boolean;
  isActive: boolean;
  isPresale: boolean;
  //   isWhitelistOnly: boolean;
  goLiveDate: number;
  goLiveDateTimeString: string;
  price: anchor.BN;
  gatekeeper: null | {
    expireOnUse: boolean;
    gatekeeperNetwork: anchor.web3.PublicKey;
  };
  endSettings: null | {
    number: anchor.BN;
    endSettingType: any;
  };
  whitelistMintSettings?: null | {
    mode: any;
    mint: anchor.web3.PublicKey;
    presale: boolean;
    discountPrice: null | anchor.BN;
  };
  hiddenSettings: null | {
    name: string;
    uri: string;
    hash: Uint8Array;
  };
  //   retainAuthority: boolean;
}

export interface ICandyMachineAccount {
  id: string;
  program: anchor.Program;
  state: ICandyMachineState;
}